# packages/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Doctor Agent Product API",
    "version": "0.3.0-mvp"
  },
  "paths": {
...
    "/api/orchestration/ask": {
      "post": {
        "tags": [
          "orchestration"
        ],
        "summary": "Ask Master",
        "description": "问主诊医生。Master 自行决定叫哪些助手、要不要采纳。\n\n**不做流式**：Master 的价值在于它对多个 worker 结果的取舍，\n那个判断只有在全部 worker 回来之后才成立。中途流式吐字会让使用者\n看到尚未被取舍的中间结论，那比等一会儿更糟。",
        "operationId": "ask_master_api_orchestration_ask_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AskIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Ask Master Api Orchestration Ask Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/orchestration/workers/{worker_name}/run": {
      "post": {
        "tags": [
          "orchestration"
        ],
        "summary": "Run Worker",
        "description": "单独跑一个 worker，绕过 Master。\n\n用途是调试与回归：要判断「是这个岗位的 skill 写坏了，还是 Master 派错了活」，\n就必须能单独驱动它。",
        "operationId": "run_worker_api_orchestration_workers__worker_name__run_post",
        "parameters": [
          {
            "name": "worker_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Worker Name"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkerRunIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Run Worker Api Orchestration Workers  Worker Name  Run Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AskIn": {
        "properties": {
          "patient_id": {
            "type": "string",
            "title": "Patient Id"
          },
          "question": {
            "type": "string",
            "title": "Question"
          },
          "workers": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Workers"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "patient_id",
          "question"
        ],
        "title": "AskIn"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "WorkerRunIn": {
        "properties": {
          "patient_id": {
            "type": "string",
            "title": "Patient Id"
          },
          "task": {
            "type": "string",
            "title": "Task"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "patient_id",
          "task"
        ],
        "title": "WorkerRunIn"
      }
    }
  }
}